make parameter ids hash based

+++ v2-swiftpm/swiftpm/Auv3PluginPackage/Sources/Dsp/definitions/api.h
@@ -5,7 +5,7 @@ class IDspCore {
 public:
   virtual ~IDspCore() = default;
   virtual void prepareProcessing(double sampleRate, uint32_t maxFrames) = 0;
-  virtual void setParameter(uint32_t id, double value) = 0;
+  virtual void setParameter(uint64_t id, double value) = 0;
   virtual void noteOn(int noteNumber, double velocity) = 0;
   virtual void noteOff(int noteNumber) = 0;
   virtual void processAudio(float *bufferL, float *bufferR,
